refactor(banners): separate health check from session watching
- HealthBanner now only polls /health and renders the unavailable banner
- SessionWatcher owns token expiry, custom refresh, and the LoggedOutModal;
  rendered as a sibling in layout.tsx
- useSessionWatcher drops healthError param (userError 403 is sufficient)
